ProjectTaskManager/pyproject.toml
[tool.poetry]
package-mode=false

[tool.poetry.dependencies]
python="^3.13.5"
fastapi = "^0.116.1"
pydantic-settings = "^2.10.1"
sqlalchemy = "^2.0.43"
uvicorn = "^0.35.0"
alembic = "^1.16.4"
asyncpg = "^0.30.0"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

